Implement weighted SmartMatch scoring engine

+++ src/PeopleHub.SmartMatch/Models/ProviderScore.cs
@@ -4,5 +4,5 @@ public sealed class ProviderScore
 {
     public Guid ProviderId { get; init; }
 
-    public int Score { get; init; }
+    public decimal Score { get; init; }
 }
